[BE] fix: 압축 소요 시간 Dto 타입 수정

See merge request s13-final/S13P31S404!101

+++ backend/src/main/java/com/semes/impresser/convertImage/dto/response/ConvertHistoryDetailResponse.java
@@ -12,7 +12,7 @@ public record ConvertHistoryDetailResponse(
     Long elapsedTime,
     String sourceExtension,
     String compressedExtension,
-    Long compressionTime
+    BigDecimal compressionTime
 ) {
 
 }
